flats_web/requirements.txt
flask
pyopenssl==22.0.0
cryptography==38.0.4
psycopg2-binary==2.9.1
